feat: build and push Docker image with GH Actions
Action to build a multi-platform Docker image from the Dockerfile in the
repo and push it to the GHCR registry.

The image is built and deployed to the GHCR registry by the
`docker-image` GH Action.

This action is triggered by a push to the `main` branch when it affects
the `Dockerfile`. It can also be triggered manually via the Actions web
UI, the GH REST API or the GH CLI tool:
```sh
gh workflow run docker-image
```

When triggered by a push, it will push to the GH registry two image tags:
  * a `version` tag where the version is the current date in ISO format
  * a `latest` tag pointing to the version tag

When triggered by hand, the workflow will run as a `dry-run` by default.
That is, it will still set the `version` tag to the current date in ISO
format but **it won't tag `latest` or push the tags to the registry**.

To override these defaults, use the `IMAGE_VERSION`, `TAG_LATEST` and
`PUSH` variables, e.g.:
```sh
gh workflow run docker-image -f IMAGE_VERSION=20241101-testing
```