feat: Dockerfile slices
Build three different "slices" (types of images): `base`, `main` and
`debug`

The GH Actions workflow will dynamically compose the Dockerfile to build
the images. It will use the `base.Dockerfile` as a "preamble" that will
be concatenated to the `<SLICE>.Dockerfile.slice` to create the final
`Dockerfile` that will be used for building.

So, each image will be built from the following Dockerfiles (in order):
  * `base`: `base.Dockerfile`
  * `main`: `base.Dockerfile`, `main.Dockerfile.slice`
  * `debug`: `base.Dockerfile`, `main.Dockerfile.slice`, `debug.Dockerfile.slice`